move user management endpoint functions to separate module
